refactor(steps): rename step files; update default dotfiles path
Remove stale numeric prefixes from all step filenames (e.g.,
07-contexts.tsx → contexts.tsx). Update all imports in wizard,
config-first, and update modes.

Change default dotfilesRepo from {context-path}/dotfiles to
{workspaceRoot}/dotfiles — one shared dotfiles repo per workspace
rather than per context.
