+++ src/modes/update.tsx
@@ -16,13 +16,13 @@ import { Box, Text, useApp } from 'ink';
 import Spinner from 'ink-spinner';
 import { loadConfig } from '../config/reader.js';
 import { atomicWriteConfig } from '../config/writer.js';
-import { ShellStep } from '../steps/02-shell.js';
-import { AppConfigStep } from '../steps/10-app-config.js';
-import { ToolsStep } from '../steps/09-tools.js';
-import { BrowserStep } from '../steps/14-browser.js';
-import { AIToolsStep } from '../steps/15-ai-tools.js';
-import { ContextsStep } from '../steps/07-contexts.js';
-import { LanguagesStep } from '../steps/05-languages.js';
+import { ShellStep } from '../steps/shell.js';
+import { AppConfigStep } from '../steps/app-config.js';
+import { ToolsStep } from '../steps/tools.js';
+import { BrowserStep } from '../steps/browser.js';
+import { AIToolsStep } from '../steps/ai-tools.js';
+import { ContextsStep } from '../steps/contexts.js';
+import { LanguagesStep } from '../steps/languages.js';
 import type { TildeConfig } from '../config/schema.js';
 import { CURRENT_SCHEMA_VERSION } from '../config/migrations/runner.js';
