+++ src/modes/wizard.tsx
@@ -3,21 +3,21 @@ import { Box, Text, useInput } from 'ink';
 import SelectInput from 'ink-select-input';
 import type { TildeConfig } from '../config/schema.js';
 import { saveCheckpoint, loadCheckpoint, clearCheckpoint } from '../state/checkpoint.js';
-import { ConfigDetectionStep } from '../steps/00-config-detection.js';
-import { EnvCaptureStep } from '../steps/01-env-capture.js';
+import { ConfigDetectionStep } from '../steps/config-detection.js';
+import { EnvCaptureStep } from '../steps/env-capture.js';
 import type { EnvironmentCaptureReport } from '../capture/scanner.js';
 import { parseGitconfig } from '../capture/parser.js';
-import { ShellStep } from '../steps/02-shell.js';
-import { PackageManagerStep } from '../steps/03-package-manager.js';
-import { VersionManagerStep } from '../steps/04-version-manager.js';
-import { ContextsStep } from '../steps/07-contexts.js';
-import { ToolsStep } from '../steps/09-tools.js';
-import { AppConfigStep } from '../steps/10-app-config.js';
-import { SecretsBackendStep } from '../steps/12-secrets-backend.js';
-import { ConfigExportStep } from '../steps/13-config-export.js';
-import { BrowserStep } from '../steps/14-browser.js';
-import { AIToolsStep } from '../steps/15-ai-tools.js';
-import { ApplyStep } from '../steps/16-apply.js';
+import { ShellStep } from '../steps/shell.js';
+import { PackageManagerStep } from '../steps/package-manager.js';
+import { VersionManagerStep } from '../steps/version-manager.js';
+import { ContextsStep } from '../steps/contexts.js';
+import { ToolsStep } from '../steps/tools.js';
+import { AppConfigStep } from '../steps/app-config.js';
+import { SecretsBackendStep } from '../steps/secrets-backend.js';
+import { ConfigExportStep } from '../steps/config-export.js';
+import { BrowserStep } from '../steps/browser.js';
+import { AIToolsStep } from '../steps/ai-tools.js';
+import { ApplyStep } from '../steps/apply.js';
 
 const DEFAULT_CONFIGURATIONS = { git: true, vscode: false, aliases: false, osDefaults: false, direnv: true };
