+++ src/steps/config-detection.tsx
